refactor(arch): move shared persistence to config/persistence
Move database.py and base.py from src/infrastructure/ to
src/config/persistence/. This eliminates the top-level infrastructure
directory, keeping src/ focused on business domain per screaming
architecture principles.
