--- src/infrastructure/__init__.py
@@ -1,7 +1,0 @@
-"""Infrastructure layer for HomeFlix.
-
-Contains shared infrastructure components:
-- Database initialization and Base model
-- External API clients (future)
-- File system services (future)
-"""
