--- src/infrastructure/persistence/models/__init__.py
@@ -1,8 +1,0 @@
-"""Shared ORM models.
-
-Contains the Base model used by all module-specific ORM models.
-"""
-
-from src.infrastructure.persistence.models.base import Base
-
-__all__ = ["Base"]
